+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove `StatsCollector` interface.
 - Remove `ClientMetricReport` interface.
+- Remove `clientMetricReport` parameter from `StatsCollector.start()` API.
 
 ### Changed